refactor: apply decomposition
