+++ src/pages/welcome/connect-to-lobby/connect-to-lobby.module.scss
@@ -1,9 +1,5 @@
 .connect {
   max-width: 876px;
-  padding-top: 60px;
-  padding-left: 30px;
-  background: #FFFFFF;
-  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
   display: flex;
 }
 
@@ -14,7 +10,6 @@
   font-size: 64px;
   line-height: 75px;
   color: #000000;
-
 }
 
 .rowTitle {
@@ -44,6 +39,7 @@
 }
 
 .container {
+  word-break: normal;
   display: flex;
   flex-direction: column;
   justify-content: space-between;
@@ -71,10 +67,10 @@
   font-size: 24px;
   line-height: 30px;
   text-align: center;
-  color: #FFFFFF;
+  color: #ffffff;
   text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-  background: #2B3A67;
-  border: 1px solid #496A81;
+  background: #2b3a67;
+  border: 1px solid #496a81;
   box-sizing: border-box;
   border-radius: 0px 3px 3px 0px;
 }
@@ -95,7 +91,7 @@
   width: 100px;
   height: 100px;
   border-radius: 50px;
-  background: #60DABF;
+  background: #60dabf;
   box-shadow: inset 0px 4px 4px rgba(0, 0, 0, 0.25);
   font-family: Roboto;
   font-style: normal;
@@ -103,31 +99,20 @@
   font-size: 48px;
   line-height: 100px;
   text-align: center;
-  color: #FFFFFF;
+  color: #ffffff;
   text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
 }
 
-.left {
-  text-align: start;
-}
-
+.left,
 .right {
   text-align: start;
-}
-
-.btnsRow {
-  margin: 100px 30px 25px 0px;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-
 }
 
 .confirm {
   width: 189px;
   height: 60px;
-  background: #2B3A67;
-  border: 1px solid #496A81;
+  background: #2b3a67;
+  border: 1px solid #496a81;
   box-sizing: border-box;
   border-radius: 3px;
   font-family: Ruda;
@@ -136,15 +121,15 @@
   font-size: 24px;
   line-height: 48px;
   text-align: center;
-  color: #FFFFFF;
+  color: #ffffff;
   text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
 }
 
 .cancel {
   width: 189px;
   height: 60px;
-  background: #FFFFFF;
-  border: 1px solid #496A81;
+  background: #ffffff;
+  border: 1px solid #496a81;
   box-sizing: border-box;
   border-radius: 3px;
   font-family: Ruda;
@@ -153,11 +138,12 @@
   font-size: 24px;
   line-height: 48px;
   text-align: center;
-  color: #2B3A67;
+  color: #2b3a67;
   text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
 }
 
 .tipError {
+  white-space: nowrap;
   display: flex;
   flex-direction: column;
   justify-content: flex-end;
@@ -175,10 +161,9 @@
   margin: 15px 0;
   text-align: center;
 }
- 
+
 .field__file {
   opacity: 0;
   visibility: hidden;
   position: absolute;
 }
-
