+++ src/pages/welcome/welcome.tsx
@@ -1,11 +1,8 @@
 import React from 'react';
 import FormWelcome from './form/form';
 
-
-const WelcomePage = ():JSX.Element => {
-  return(
-      <FormWelcome/>
-  )
-}
+const WelcomePage = (): JSX.Element => {
+  return <FormWelcome />;
+};
 
 export default WelcomePage;
