+++ src/redux/types/connect-to-lobby.ts
@@ -1,0 +1,4 @@
+export interface IUseFormProps {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  reg: any;
+}
